Debugger changes.
Creating a full debugger GUI is just too much of an investment right
now to warrant persuing it.  I would much rather have a gdb-style
debugger that is much easier to implement than the original. I'm
not removing the dbg-gui/dbg-pane code from the repo in case I want
to revisit it, but safe to say it's not actively in the works. I did
update the Makefile such that "make dbg-gui" will build it.