Cartridge interface with Mapper 0 iNES Cartridge implemented.
Among other minor fixes, the cartridge interface (icart.h) was
defined, along with the creation of the iNES cartridge parser,
which is a definite work in progress.  Incorporated the cartridge
loader with the main program, and (for testing purposes) have forced
the initial PC of the CPU to 0xC000.  Added a target for testing the
cartridge loader.